ewetzha: compiled with ++x14 instead of 11, turned on -Wall (all warnings) + fix a nunber of warbnings + added flaf Werror which turnes` warnings into errors (occi code is not compliant) + controll WError from setnev file

+++ src/cpp/experiments/algorithms/selection/track-m-smallest-numbers/test1.cc
@@ -32,7 +32,7 @@ public:
   // add a number
   void add(int val){
     q_.push(val);
-    if(q_.size()>m_)q_.pop();
+    if(static_cast<int>(q_.size())>m_)q_.pop();
   }
   // get reference to vector containing smallest numbers so far
   vector<int>get()const{
